cart app created

+++ e-commerce/online_shopping/products/models.py
@@ -11,7 +11,7 @@ class Product(models.Model):
     price = models.DecimalField(max_digits=10, decimal_places=2)
     categories = models.ManyToManyField(Category, related_name='products')
     brand = models.CharField(max_length=50)
-    rating = models.IntegerField()
+    rating = models.BigIntegerField()
     availability = models.BooleanField()
     
     def __str__(self):
